requirements.txt
click
psycopg2-binary
jinja2
cerberus
pyyaml
argparse